refactor(MetricUsageFetcher): Split code into several files
