+++ src/WingmanDataTrail/ListControls/MetricsSorter/fetchers/extractMetricNames.ts
@@ -1,0 +1,31 @@
+import { parser } from '@prometheus-io/lezer-promql';
+
+/**
+ * Extracts all metric names from a PromQL expression
+ * @param {string} promqlExpression - The PromQL expression to parse
+ * @returns {string[]} An array of unique metric names found in the expression
+ */
+export function extractMetricNames(promqlExpression: string): string[] {
+  const tree = parser.parse(promqlExpression);
+  const metricNames = new Set<string>();
+  const cursor = tree.cursor();
+
+  do {
+    // when we find a VectorSelector...
+    if (cursor.type.is('VectorSelector')) {
+      // go to its first child
+      if (cursor.firstChild()) {
+        do {
+          // look for the Identifier node
+          if (cursor.type.is('Identifier')) {
+            const metricName = promqlExpression.slice(cursor.from, cursor.to);
+            metricNames.add(metricName);
+          }
+        } while (cursor.nextSibling());
+        cursor.parent();
+      }
+    }
+  } while (cursor.next());
+
+  return Array.from(metricNames);
+}
